MQTT, OTA changes

+++ src/cmnds/cmd_if.c
@@ -1,4 +1,3 @@
-
 #include "../new_common.h"
 #include "cmd_local.h"
 #include "../httpserver/new_http.h"
@@ -155,7 +154,7 @@ float CMD_EvaluateExpression(const char *s, const char *stop) {
 	if(stop == 0) {
 		stop = s + strlen(s);
 	}
-	while(stop > s && isspace(stop[-1])) {
+	while(stop > s && isspace(((int)stop[-1]))) {
 		stop --;
 	}
 	if(g_expDebugBuffer==0){
